feat(makesky): change binning to order used in driftscan
This changes the frequency specification such that the binning of
channels is performed before the selection. This matches what is done in
driftscan.  This would be a breaking change if the binning option was
more frequently used.

This also changes the frequency specification to more naturally match
what's done for CHIME (i.e. skipping the Nyquist frequency) and a new
`--freq-mode=centre_nyquist` option to select the previous behaviour.