BREAKING CHANGE : Introduces `tracer' command.
The tracer command accepts a `ttype' option which, for now, can be
either 'qso' or 'none' (default). In the future, other tracers will
be added.

This commit also adds the option `bias' for the 21cm command which can
be used with or without the ZA option. It accepts a positive number
for constant bias or 0 to set the internal calculation based on the Halo Model.