fix(skysim, nputil): fix internal convention in matrix_root_manynull
matrix_root_manynull constructs a matrix square root using an eigenvalue decomposition,
but there is an ambiguity in the result stemming from the arbitrary signs of the
eigenvectors used in this construction. When this is used for generating correlated
Gaussian realizations, this can lead to inconsistent results even for the same initial
seed (e.g. for two covariances matrices which differ only by a constant multiplier,
the derived realizations can be different even though they should have correlation
coefficient equal to unity).

This commit implements a consistent convention within matrix_root_manynull(), by forcing
the first component of each eigenvector to be positive. This is set as the default,
and also explicitly set in skysim.mkfullsky(), which is the main context where this matters.